requirements.txt
django==2.2
djangorestframework==3.12.1
markdown==3.2.1
django-filter==2.4.0
mysqlclient==2.0.1
djangorestframework-simplejwt==4.4.0
django-phonenumber-field==5.0.0
phonenumbers
Pillow